--- internal/ctxkey/ctxkey.go
@@ -1,36 +1,0 @@
-// Package backoff provides context key function
-package ctxkey
-
-import "context"
-
-type contextKey string
-
-const grpcMethodContextKey contextKey = "grpc_method"
-
-// WithGRPCMethod returns a copy of parent in which the method associated with key (grpcMethodContextKey).
-func WithGRPCMethod(ctx context.Context, method string) context.Context {
-	return context.WithValue(ctx, grpcMethodContextKey, method)
-}
-
-// FromGRPCMethod returns the value associated with this context for key (grpcMethodContextKey).
-func FromGRPCMethod(ctx context.Context) string {
-	if v := ctx.Value(grpcMethodContextKey); v != nil {
-		return v.(string)
-	}
-	return ""
-}
-
-const backoffNameContextKey contextKey = "backoff_name"
-
-// WithBackoffName returns a copy of parent in which the method associated with key (backoffNameContextKey).
-func WithBackoffName(ctx context.Context, name string) context.Context {
-	return context.WithValue(ctx, backoffNameContextKey, name)
-}
-
-// FromBackoffName returns the value associated with this context for key (backoffNameContextKey).
-func FromBackoffName(ctx context.Context) string {
-	if val := ctx.Value(backoffNameContextKey); val != nil {
-		return val.(string)
-	}
-	return ""
-}
