+++ internal/net/grpc/client.go
@@ -25,7 +25,6 @@ import (
 	"time"
 
 	"github.com/vdaas/vald/internal/backoff"
-	"github.com/vdaas/vald/internal/ctxkey"
 	"github.com/vdaas/vald/internal/errgroup"
 	"github.com/vdaas/vald/internal/errors"
 	"github.com/vdaas/vald/internal/log"
@@ -551,8 +550,8 @@ func (g *gRPCClient) do(ctx context.Context, p pool.Conn, addr string, enableBac
 		}
 	}()
 	if g.bo != nil && enableBackoff {
-		if method := ctxkey.FromGRPCMethod(sctx); len(method) != 0 {
-			sctx = ctxkey.WithBackoffName(ctx, method+"/"+addr)
+		if method := FromGRPCMethod(sctx); len(method) != 0 {
+			sctx = backoff.WithBackoffName(ctx, method+"/"+addr)
 		}
 		data, err = g.bo.Do(sctx, func(ictx context.Context) (r interface{}, ret bool, err error) {
 			err = p.Do(func(conn *ClientConn) (err error) {
